# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ethnos.app Academic Bibliography API",
    "version": "2.0.0",
    "description": "\nREST API for academic bibliography: search, works, persons, institutions, venues, courses, citations, collaborations and health.\n\nHighlights\n- Public endpoints; selected admin/metrics routes require 'x-access-key'.\n- Standard response envelope: { status, data, pagination?, meta? }.\n- Pagination supports page/limit and offset/limit.\n- Identifiers: DOI, ORCID, ROR, ISSN.\n\nNotes\n- Rate limiting is configurable via environment and enabled by default.\n- Performance and dataset size vary by deployment.\n        ",
    "contact": {
      "name": "Bruno Cesar Cunha Cruz, PhD Student"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-developer-orcid": "0000-0001-8652-2333",
    "x-institution": "PPGAS/MN/UFRJ"
  },
  "servers": [
    {
      "url": "https://api.ethnos.app",
      "description": "Production API Server - Enterprise-ready academic research infrastructure"
    },
    {
      "url": "http://localhost:1210",
      "description": "Development Server - Local testing and development environment"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "XAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-key",
        "description": "Internal access key required for protected endpoints (security, dashboard, health except /liveness)"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "Response payload. Varies per endpoint."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 5952,
            "description": "Unique identifier for the person"
          },
          "preferred_name": {
            "type": "string",
            "example": "Dr. Maria Silva Santos",
            "description": "Preferred display name of the person"
          },
          "given_names": {
            "type": "string",
            "example": "Maria Silva",
            "description": "Given names"
          },
          "family_name": {
            "type": "string",
            "example": "Santos",
            "description": "Family name or surname"
          },
          "orcid": {
            "type": "string",
            "example": "0000-0002-1825-0097",
            "description": "ORCID identifier"
          },
          "lattes_id": {
            "type": "string",
            "example": "1234567890123456",
            "description": "Lattes CV platform ID (Brazil)"
          },
          "scopus_id": {
            "type": "string",
            "example": "57194582100",
            "description": "Scopus Author ID"
          },
          "wos_id": {
            "type": "string",
            "example": "A-1234-2023",
            "description": "Web of Science ResearcherID"
          },
          "primary_affiliation": {
            "$ref": "#/components/schemas/Organization",
            "description": "Primary institutional affiliation"
          },
          "works_count": {
            "type": "integer",
            "example": 45,
            "description": "Total number of works authored"
          },
          "h_index": {
            "type": "integer",
            "example": 12,
            "description": "H-index bibliometric indicator"
          },
          "citation_count": {
            "type": "integer",
            "example": 678,
            "description": "Total citations received"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 12345,
            "description": "Unique identifier for the organization"
          },
          "name": {
            "type": "string",
            "example": "Universidade de São Paulo",
            "description": "Full name of the organization"
          },
          "short_name": {
            "type": "string",
            "example": "USP",
            "description": "Short name or acronym"
          },
          "type": {
            "type": "string",
            "enum": [
              "UNIVERSITY",
              "RESEARCH_INSTITUTE",
              "COMPANY",
              "GOVERNMENT",
              "NGO",
              "HOSPITAL"
            ],
            "example": "UNIVERSITY",
            "description": "Type of organization"
          },
          "country": {
            "type": "string",
            "example": "Brazil",
            "description": "Country where organization is located"
          },
          "region": {
            "type": "string",
            "example": "South America",
            "description": "Geographic region"
          },
          "city": {
            "type": "string",
            "example": "São Paulo",
            "description": "City location"
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://www.usp.br",
            "description": "Official website URL"
          },
          "ror_id": {
            "type": "string",
            "example": "https://ror.org/036rp1748",
            "description": "Research Organization Registry ID"
          },
          "works_count": {
            "type": "integer",
            "example": 125420,
            "description": "Total number of works from this organization"
          },
          "members_count": {
            "type": "integer",
            "example": 8250,
            "description": "Number of affiliated researchers"
          },
          "h_index": {
            "type": "integer",
            "example": 245,
            "description": "Institutional H-index"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 6894,
            "description": "Total number of results"
          },
          "page": {
            "type": "integer",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "example": 20,
            "description": "Number of results per page"
          },
          "totalPages": {
            "type": "integer",
            "example": 345,
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "example": true,
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "example": false,
            "description": "Whether there is a previous page"
          }
        }
      },
...
    },
    "parameters": {
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Number of items to return per page (works with both page and offset formats)",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "offsetParam": {
        "name": "offset",
        "in": "query",
        "description": "Number of items to skip. Automatically converts to page format: offset÷limit = page number",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        },
        "example": 20
      },
      "pageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based). Use with limit parameter for traditional pagination",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "example": 2
      },
...
    },
    "responses": {
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request - Invalid input parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Invalid venue ID",
              "code": "VALIDATION_ERROR"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Resource not found",
              "code": "NOT_FOUND"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/bibliographies": {
      "get": {
        "summary": "List bibliography entries",
        "tags": [
          "Bibliography"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "in": "query",
            "name": "course_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "work_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "instructor_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reading_type",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "week_number",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_from",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "program_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "light",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
    "/bibliographies/analyses": {
      "get": {
        "summary": "Bibliography analysis",
        "tags": [
          "Bibliography"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "year_from",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "program_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reading_type",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/signatures/statistics": {
      "get": {
        "summary": "Get signature statistics",
        "description": "Retrieve comprehensive statistics about signatures in the system",
        "tags": [
          "Signatures"
        ],
        "responses": {
          "200": {
            "description": "Signature statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_signatures": {
                          "type": "integer",
                          "example": 378134
                        },
                        "average_signature_length": {
                          "type": "number",
                          "format": "float",
                          "example": 10.17
                        },
                        "total_characters": {
                          "type": "integer",
                          "example": 3847823
                        },
                        "shortest_signature": {
                          "type": "integer",
                          "example": 1
                        },
                        "longest_signature": {
                          "type": "integer",
                          "example": 255
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/signatures/search": {
      "get": {
        "summary": "Search signatures",
        "description": "Search for signatures by name with exact and fuzzy matching options",
        "tags": [
          "Signatures"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "description": "Search query for signature name",
            "example": "J. Smith"
          },
          {
            "in": "query",
            "name": "exact",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Whether to perform exact matching"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Signature"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "query": {
                      "type": "string",
                      "example": "J. Smith"
                    },
                    "exact_match": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/signatures/{id}": {
      "get": {
        "summary": "Get signature by ID",
        "description": "Retrieve detailed information about a specific signature",
        "tags": [
          "Signatures"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Signature ID",
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Signature details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Signature"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/signatures/{id}/persons": {
      "get": {
        "summary": "Get persons associated with a signature",
        "description": "Retrieve all persons associated with a specific signature",
        "tags": [
          "Signatures"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Signature ID",
            "example": 123
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Associated persons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Person"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}